refactor: move orchestration controller and view to collavre engine
- Move OrchestrationController from host app to Collavre::Admin::OrchestrationController
- Move orchestration view to engine views folder
- Update routes to use engine controller
- Tests remain in host app (integration tests for host app routes)
